+++ src/2023/Talks/UseFetchTalks.ts
@@ -1,6 +1,7 @@
 import {useQuery, UseQueryResult} from "react-query";
-import {IGroup, Session,} from "./Talk.types";
+
 import axios from "axios";
+import {IGroup, Session} from "../../types/sessions";
 
 export const useFetchTalks = (): UseQueryResult<IGroup[]> =>
   useQuery("api-talks", async () => {
