+++ src/views/MeetingDetail/TalkDetailContainer2024.tsx
@@ -8,10 +8,10 @@ import conferenceData from "../../data/2024.json";
 import {useFetchTalksById} from "../Talks/UseFetchTalks";
 import * as Sentry from "@sentry/react";
 import {useFetchSpeakers} from "../Speakers/UseFetchSpeakers";
-import {Session} from "../Talks/Talk.types";
 import MeetingDetail from "./MeetingDetail";
 import {ISpeaker} from "../../types/speakers";
 import {sessionAdapter} from "../../services/sessionsAdapter";
+import {Session} from "../../types/sessions";
 
 const StyledContainer = styled.div`
   background-color: ${Color.WHITE};
